% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootCases.R
\name{bootCases}
\alias{bootCases}
\title{Empirical boostrap for linear regression coefficients}
\usage{
bootCases(formula, data, B)
}
\arguments{
\item{formula}{A string that can be coerced into class 'formula'. Usually of the form
response variable \eqn{\sim} predictor variables.A symbolic description of the model to be fitted.}

\item{data}{data frame, list or environment (or object coercible by as.data.frame to a data frame)
containing the variables in the model.}

\item{B}{The number of bootstrap replicates. Usually this will be a single positive integer.}
}
\value{
A B x (k+1) data frame containing B samples from the distributions of each of the (k+1) model
parameters, where k is the number of predictors in the model.
}
\description{
Empirical boostrap for linear regression coefficients (Description)
}
\details{
This function simulates the sampling distribution of simple linear regression coefficients
by first simulating the joint distribution of (\eqn{\mathbf{X},Y}) by the empirical distribution
function of the cases (i.e. resampling cases with replacement).
}
\references{
Davison, A.C. and Hinkley, D.V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.
}
\author{
Natalie DelRocco
}
