% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootMSE.R
\name{bootMSE}
\alias{bootMSE}
\title{MSE for bootstrap samples of linear regression coefficients}
\usage{
bootMSE(B, lmodObs, x, ncovs)
}
\arguments{
\item{B}{The number of bootstrap replicates. Usually this will be a single positive integer.}

\item{lmodObs}{The observed linear model estimated by least squares. A fitted model object of
class inheriting from 'lm'.}

\item{x}{A B x (k+1) data frame containing B samples from the distributions of each of the (k+1) model
parameters, where k is the number of predictors in the model.}

\item{ncovs}{The number of covariates in the observed model. A positive integer.}
}
\value{
A vector of size (k+1) containing the mean square error of the bootstrap sample of
each regression coefficient.
}
\description{
MSE of a bootstrap sample of simple linear regression coefficients
}
\details{
This function calculates the mean square error for bootstrap samples of linear regression coefficients.
}
\author{
Natalie DelRocco
}
