% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapSamples.R
\name{bootstrapSamples}
\alias{bootstrapSamples}
\title{Obtain bootstrap samples for linear regression coefficients}
\usage{
bootstrapSamples(lmodObs, formula, data, B)
}
\arguments{
\item{lmodObs}{The observed linear model estimated by least squares. A fitted model object of
class inheriting from 'lm'.}

\item{formula}{A string that can be coerced into class 'formula'. Usually of the form
response variable \eqn{\sim} predictor variables. A symbolic description of the model to be fitted.}

\item{data}{data frame, list or environment (or object coercible by as.data.frame to a data frame)
containing the variables in the model.}

\item{B}{The number of bootstrap replicates. Usually this will be a single positive integer.}
}
\description{
Obtain bootstrap samples for linear regression coefficients using six different methods.
}
\details{
This function is a wrapper for bootCases, bootResids, and bootWild.
This function obtains the bootstrapped sampling distributions for linear
regression coefficients obtained in each of six methods: case resampling, residual resampling,
Wild bootstrap with Mammen's Two-Point Distribution, Wild bootstrap with Mammen's Continuous
Distribution, Wild boostrap with Rademacher Distribution, and Wild bootstrap with Standard Normal
Distribution.
}
\author{
Natalie DelRocco
}
