% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapSummary.R
\name{bootstrapSummary}
\alias{bootstrapSummary}
\title{Compare bootstrapping methods for linear regression coefficients using summary statistics.}
\usage{
bootstrapSummary(bootList, lmodObs, summaryStats = c("mean", "median",
  "Q1", "Q3", "var", "IQR", "se", "MSE"))
}
\arguments{
\item{bootList}{A list containing the bootstrapped sampling distributions of linear regression
coefficients. Each element in the list should contain a data frame obtained using one of the
six supported bootstrapping methods of size B x (k+1) where k is the number of predictors in the
model. Column \eqn{i} of the data frame will be a sample of size B from the sampling distribution of
\eqn{\beta_{i}}.}

\item{lmodObs}{The observed linear model estimated by least squares. A fitted model object of
class inheriting from 'lm'.}

\item{summaryStats}{A string or a vector of strings. The desired summary statistics of the sampling
of regression coefficients for the function to return. Options are: mean, median, Q1, Q3, IQR, var,
se, and MSE.}
}
\description{
Compare bootstrapping methods for linear regression coefficients using summary statistics.
}
\details{
This function compares the summary statistics of bootstrapped sampling distributions for linear
regression coefficients obtained in each of six methods: case resampling, residual resampling,
Wild bootstrap with Mammen's Two-Point Distribution, Wild bootstrap with Mammen's Continuous
Distribution, Wild boostrap with Rademacher Distribution, and Wild bootstrap with Standard Normal
Distribution.
}
\examples{
library(faraway)
data(prostate)
lmod <- lm(lpsa ~ lcavol + age + lweight, prostate)
bootsamps <- bootstrapSamples(lmodObs=lmod, formula='lpsa ~ lcavol + age + lweight',
                              data=prostate, B=1000)
bootstrapSummary(bootList = bootsamps, lmodObs = lmod,
                 summaryStats = c('median', 'var', 'se', 'IQR'))

}
\author{
Natalie DelRocco
}
