% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficientPlots.R
\name{coefficientPlots}
\alias{coefficientPlots}
\title{Visualize bootstrapped sampling distribution for linear regression coefficients}
\usage{
coefficientPlots(lmodObs, bootList)
}
\arguments{
\item{lmodObs}{The observed linear model estimated by least squares. A fitted model object of
class inheriting from 'lm'.}

\item{bootList}{A list containing the bootstrapped sampling distributions of linear regression
coefficients. Each element in the list should contain a data frame obtained using one of the
six supported bootstrapping methods of size B x (k+1) where k is the number of predictors in the
model. Column \eqn{i} of the data frame will be a sample of size B from the sampling distribution of
\eqn{\beta_{i}}.}
}
\description{
Visualize the bootstrapped sampling distribution for linear regression coefficients (Description)
}
\details{
This function plots the bootstrapped sampling distributions of linear regression coefficients obtained
using six different methods: case resampling, residual resampling,
Wild bootstrap with Mammen's Two-Point Distribution, Wild bootstrap with Mammen's Continuous
Distribution, Wild boostrap with Rademacher Distribution, and Wild bootstrap with Standard Normal
Distribution. Each plot contains (k+1) facets where k is the number of predictors in the model. The
least squares estimate of each regression coefficient is denoted with a red dashed line.
}
\author{
Natalie DelRocco
}
