% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMammen2.R
\name{rMammen2}
\alias{rMammen2}
\title{Random number generation for Mammen's two-point distribution}
\usage{
rMammen2(n)
}
\arguments{
\item{n}{A positive integer. The number of random deviates to generate.}
}
\value{
n random realizations from Mammen's two-point distribution
}
\description{
Generate a random deviate from Mammen's two-point distribution (Description)
}
\details{
This function generates random numbers from Mammen's two-point distribution:
}
\examples{
rMammen2(1)

}
\references{
Davison, A.C. and Hinkley, D.V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.

Mammen, E. (1993). Bootstrap and Wild Bootstrap for High Dimensional Linear Models. The Annals of Statistics, 21(1), 255-285. doi: 10.1214/aos/1176349025
}
\author{
Natalie DelRocco
}
