% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMammenC.R
\name{rMammenC}
\alias{rMammenC}
\title{Random number generation for Mammen's continuous distribution}
\usage{
rMammenC(n)
}
\arguments{
\item{n}{A positive integer. The number of random deviates to generate.}
}
\value{
n random realizations from Mammen's continuous distribution
}
\description{
Generate random deviates from Mammen's continuous distribution (Description)
}
\details{
This function generates random numbers from Mammen's continuous distribution.
}
\references{
Davison, A.C. and Hinkley, D.V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.

Mammen, E. (1993). Bootstrap and Wild Bootstrap for High Dimensional Linear Models. The Annals of Statistics, 21(1), 255-285. doi: 10.1214/aos/1176349025
}
\author{
Natalie DelRocco
}
