% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRademacher.R
\name{rRademacher}
\alias{rRademacher}
\title{Random number generation for the Rademacher distribution}
\usage{
rRademacher(n)
}
\arguments{
\item{n}{A positive integer. The number of random deviates to generate.}
}
\value{
n random realizations from the Rademacher distribution
}
\description{
Generate random deviates from the Rademacher distribution (Description)
}
\details{
This function generates random numbers from the Rademacher distribution.
}
\author{
Natalie DelRocco
}
