% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler}
\alias{sampler}
\title{Obtain bootstrap samples of an input dataset.}
\usage{
sampler(x, B)
}
\arguments{
\item{x}{An input dataset.}

\item{B}{The number of bootstrap replicates. Usually this will be a single positive integer.}
}
\description{
Obtain bootstrap samples of an input dataset.
}
\details{
This a generic function for resampling with replacement and equal probability from an input dataset.
}
\author{
Natalie DelRocco
}
